Add RSC prefetching to uncategorized page to eliminate duplicate entry fetch
The uncategorized page was a pure client component without server-side
prefetching. When opening an entry, this caused a second network request
to entries.get after the RSC navigation. Converted to server/client
component pattern matching other pages (all, saved, starred) to prefetch
entry content, entries list, and subscriptions on the server.
